Atualiza projeto

+++ backend/car-catalog-backend/src/main/java/com/wswork/carcatalog/dto/CarroFormatadoDTO.java
@@ -4,8 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 /**
  * DTO para retornar carros no formato específico do cars.json
  */
@@ -22,6 +20,6 @@ public class CarroFormatadoDTO {
     private Integer num_portas;
     private String cor;
     private String nome_modelo;
-    private BigDecimal valor;
+    private Integer valor; // Alterado para Integer para ficar igual ao exemplo
     private String marca; // Nome da marca conforme especificação do teste
 }
